pdf-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
cloudinary